% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_emissions.R
\name{calculate_emissions}
\alias{calculate_emissions}
\title{Get Hourly Emissions}
\usage{
calculate_emissions(df = atref, eic)
}
\arguments{
\item{df}{Cross-referencing table}

\item{eic}{EIC code}
}
\description{
Input an EIC code and obtain a data frame (within a list) of estimated hourly emissions, by month and day of week
}
\details{
Note: only works when the four counties use the same profile codes for a given EIC (i.e. the four counties have to have the same temporal profile)
}
