% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_table.R
\name{get_code_table}
\alias{get_code_table}
\title{Get Code Table}
\usage{
get_code_table(df = atref, eic)
}
\arguments{
\item{df}{Cross-referencing table}

\item{eic}{EIC code}
}
\description{
Input an EIC code and obtain the list of profile codes in the South Coast Basin
}
