% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_ratios.R
\name{get_profile_ratios}
\alias{get_profile_ratios}
\title{Get Temporal Profiles}
\usage{
get_profile_ratios(df = atref, eic)
}
\arguments{
\item{df}{Cross-referencing table}

\item{eic}{EIC code}
}
\description{
Input an EIC code and obtain list of hourly, day of week, and monthly temporal profiles (ratios)
}
\details{
Note: only works when the four counties use the same profile codes for a given EIC (i.e. the four counties have to have the same temporal profile)
}
