% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_code.R
\name{get_time_codes}
\alias{get_time_codes}
\title{Get Time Codes}
\usage{
get_time_codes(df = atref, eic)
}
\arguments{
\item{df}{Cross-referencing table}

\item{eic}{EIC code}
}
\description{
Input an EIC code and obtain a named vector of profile codes in the South Coast Basin
}
\details{
Note: only works when the four counties use the same profile codes for a given EIC (i.e. the four counties have to have the same temporal profile)
}
