% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emissions.R
\name{plot_emissions}
\alias{plot_emissions}
\title{Plot Hourly Emissions}
\usage{
plot_emissions(df = atref, eic, dow)
}
\arguments{
\item{df}{Cross-referencing table}

\item{eic}{EIC code}

\item{dow}{Day of the week}
}
\description{
Input an EIC code and obtain a facet plot of estimated hourly emissions, by month and day of week
}
\details{
Note: only works when the four counties use the same profile codes for a given EIC (i.e. the four counties have to have the same temporal profile). Can only plot one day of the week at a time.
}
